카드 상세 화면 로직 구현

+++ data/paging/src/main/java/com/phew/paging/PagerRepositoryImpl.kt
@@ -29,7 +29,7 @@ class PagerRepositoryImpl @Inject constructor(
     ) { pagingUnReadNotificationProvider.get() }.flow
 
     override fun cardComments(
-        cardId: Int,
+        cardId: Long,
         latitude: Double?,
         longitude: Double?
     ): Flow<PagingData<CardComment>> =
